Rating system to be done but im too hyped to finish this today

+++ DinnerHall/Data/DistributionData.cs
@@ -12,7 +12,7 @@ namespace DinnerHall
         public int[] items { get; set; }
         public int priority { get; set; }
         public int max_wait { get; set; }
-        public int pick_up_time { get; set; }
+        public long pick_up_time { get; set; }
         public int cooking_time { get; set; }
         public Dictionary<string, int>[] cooking_details { get; set; }
